refactor(installations): split versions.ts into store + sync engine (move 2)
Move path/enumeration helpers (getVersionDir, getVersionHomePath,
listInstalledVersions, isVersionIsolated, getEffectiveHome, and the
helpers they need) into installations/store.ts so plugins/rules/
agent-spec/devices can import the store half without pulling the
resource-sync engine. versions.ts re-exports the store so existing
deep imports keep working.

getAvailableResources stays in versions.ts because it calls
discoverPlugins. fleet-inventory.ts therefore still imports versions.ts.

Prune barrel exports that have zero importers outside installations/;
keep createInstallation and recordRelease (used by commands/update.test.ts).
