+++ apps/cli/src/lib/devices/harness-inventory.ts
@@ -29,7 +29,7 @@ import {
   type UsageIdentityInput,
   type UsageSnapshot,
 } from '../accounting/usage.js';
-import { getVersionHomePath, listInstalledVersions } from '../installations/versions.js';
+import { getVersionHomePath, listInstalledVersions } from '../installations/store.js';
 import { readMeta } from '../state.js';
 import { listNativeAccounts } from '../account-registry.js';
