+++ apps/cli/src/lib/plugins/skills.ts
@@ -16,7 +16,7 @@ import { normalizeAliases } from '../resource-aliases.js';
 import { AGENTS, ensureSkillsDir, agentConfigDirName } from '../agents.js';
 import { capableAgents, isCapable } from '../capabilities.js';
 import { getAgentsDir, getUserSkillsDir, getSkillsDir as getSystemSkillsDir, getProjectAgentsDir, getEnabledExtraRepos, getTrashSkillsDir } from '../state.js';
-import { getEffectiveHome, getVersionHomePath, listInstalledVersions } from '../installations/versions.js';
+import { getEffectiveHome, getVersionHomePath, listInstalledVersions } from '../installations/store.js';
 import { listCommandSkillsInVersion } from '../command-skills.js';
 import { emit } from '../feed/events.js';
